server/requirements.txt
black==22.8.0
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Mail==0.9.1
Flask-QRcode==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.3
pandas==1.5.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
qrcode==7.3.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
